separated ProductSelect component from ProductInfo for implementing adding to bag feature on a specific scope
