+++ src/components/product_info/index.jsx
@@ -1,20 +1,10 @@
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 
-import SizeChartDrawer from '@/components/size_chart_drawer/index.jsx';
-import SizeChartTable from '@/components/size_chart_drawer/size_chart_table/index.jsx';
-import SizeSelector from '@/components/size_selector/index.jsx';
+import ProductSelect from '@/components/product_select/index.jsx';
 
 import '@/components/product_info/index.scss';
 
 function ProductInfo({ product }) {
-  // Size Chart
-  const sizeChartRef = useRef();
-  const handleSizeGuideClick = () => {
-    if (sizeChartRef.current) {
-      sizeChartRef.current.open();
-    }
-  };
-
   // Product Details
   const [showDetails, setShowDetails] = useState(false);
 
@@ -24,30 +14,7 @@ function ProductInfo({ product }) {
   return (
     <div className="product-info-wrapper">
       <div className="product-info">
-        <div className="product-metadata">
-          <div className="product-name">{product.name}</div>
-          <div className="product-price">${product.price}</div>
-        </div>
-        <div className="product-order-form">
-          <div className="product-color">Color: {product.color}</div>
-          <div className="size-chart-click-wrapper">
-            <div
-              className="size-chart-click normal-link"
-              onClick={handleSizeGuideClick}
-            >
-              Size Chart
-            </div>
-            <SizeChartDrawer ref={sizeChartRef}>
-              <SizeChartTable />
-            </SizeChartDrawer>
-          </div>
-          {/* TODO: Implement Select Size function */}
-          <div className="size-select-button-wrapper">
-            <SizeSelector sizes={product.sizes} />
-          </div>
-        </div>
-        {/* TODO: Implement Buy Button function */}
-        <button className="buy-button">Add to Shopping Bag</button>
+        <ProductSelect product={product} />
         <div className="product-description-wrapper">
           <div className="product-description">{product.description}</div>
         </div>
