+++ src/components/product_select/index.scss
@@ -1,0 +1,45 @@
+.product-select {
+  .product-metadata {
+    margin-bottom: 35px;
+
+    .product-name {
+      margin-top: 2px;
+      margin-bottom: 6px;
+    }
+
+    .product-price {
+      margin-top: 2px;
+      margin-bottom: 6px;
+    }
+  }
+  .product-order-form {
+    margin-bottom: 15px;
+
+    .size-chart-click-wrapper {
+      text-align: right;
+      margin-bottom: 10px;
+
+      .size-chart-click {
+        display: inline-block;
+        padding-bottom: 2px;
+        border-bottom: 1px solid transparent;
+      }
+    }
+    .size-select-button-wrapper {
+      width: 100%;
+    }
+  }
+  .buy-button {
+    width: 100%;
+    height: 35px;
+    text-align: center;
+    background-color: rgb(33, 33, 33);
+    color: white;
+
+    margin-bottom: 25px;
+
+    &:hover {
+      background-color: rgb(70, 70, 70);
+    }
+  }
+}
